+++ vscode-extension/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "emacs-themes" extension.
 
+## 0.0.2 (2025-11-15)
+
+- Add input/inlineChatInput colors
+- Add status bar colors
+
 ## 0.0.1 (2025-11-09)
 
 - Initial release